Move the Program into a new package, implementing Junit Testing
