+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
+    id 'org.springframework.boot' version '2.5.6'
+    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'java'
-    id 'application'
 }
 
 repositories {
@@ -8,24 +9,20 @@ repositories {
 }
 
 dependencies {
-    implementation('org.springframework.boot:spring-boot-starter-web:2.5.6')
-    implementation('org.springframework.boot:spring-boot-starter-data-jpa:2.5.6')
-    implementation 'junit:junit:4.13.2'
-    testImplementation('org.junit.jupiter:junit-jupiter:5.8.1')
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     implementation group: 'org.eclipse.persistence', name: 'javax.persistence', version: '2.2.1'
     implementation 'org.sat4j:org.sat4j.core:2.3.1'
     implementation 'org.sat4j:org.sat4j.pb:2.3.1'
     testImplementation 'org.mockito:mockito-core:4.0.0'
+    runtimeOnly 'com.h2database:h2'
+    testCompileOnly 'junit:junit:4.12'
 }
 
-mainClassName = 'Dummy'
-
 test {
+    useJUnitPlatform()
     testLogging {
         events "passed", "skipped", "failed"
     }
-}
-
-run {
-    enableAssertions = true
 }
