+++ crcCards/EmployeeManager.txt
@@ -3,7 +3,7 @@ Parent class: None
 Subclass: None
 
 Responsibility:
-* controller that enables HR access Employee Modifier to hire and fire employees
+* com.hr.controller that enables HR access Employee Modifier to hire and fire employees
 
 Collaborator：
