+++ src/main/java/com/hr/Application.java
@@ -1,4 +1,4 @@
-package service;
+package com.hr;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
