+++ src/main/java/com/hr/controller/WelcomeController.java
@@ -1,4 +1,4 @@
-package controller;
+package com.hr.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
